This is a little code to fetch data about your Facebook posts. It gives you: the post itself, the date and time it was created, the type of the message, count of shares, info about likes and comments (only showing their total count), and the unique ID of the post. Then, it converts this data into a csv file.
Troubles and issues:
No header - have to add that in Excel/Sheets/LibreCalc...
When there is an empty field (no message - which happens when it is a shared
post without any writing added to it; no shares etc.) it does not leave a
blank field. It simply writes the data as they go into the csv. So the csv
needs to be cleaned - data moved and fields emptied.
??? Found any others? I have not so far...